+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 
 - Bug fixed
 - Edit Readme
+
+## 06/19/2021
+
+- Add `editable` parameter